refactor: ViewToggle로 정리하기
